Move LocationHandler into 'handlers' package
